requirement.txt
pynput
pyperclip
pillow
cryptography
yagmail
